WIP: T12 -- Split handlers-tab.test.tsx
