+++ design/specs/016-handler-telemetry-debt/tasks/T12-split-handlers-tab-tests.md
@@ -1,7 +1,7 @@
 ---
 task_id: "T12"
 title: "Split handlers-tab.test.tsx"
-status: "planned"
+status: "done"
 depends_on: ["T10"]
 implements: ["FR#11", "AC#3"]
 ---
