--- frontend/src/components/app-detail/handlers-tab.test.tsx
@@ -1,725 +1,0 @@
-import { signal } from "@preact/signals";
-import { fireEvent, waitFor } from "@testing-library/preact";
-import { http, HttpResponse } from "msw";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-
-import { createJob, createListener } from "../../test/factories";
-import { createWouterMock } from "../../test/mock-wouter";
-import { renderWithAppState } from "../../test/render-helpers";
-import { server } from "../../test/server";
-import { HandlersTab } from "./handlers-tab";
-
-// Mock child components that make API calls
-vi.mock("../shared/execution-table", () => ({
-  ExecutionTable: ({ tableId, kind, records }: { tableId: string; kind: string; records: unknown[] }) => (
-    <div data-testid={tableId} data-kind={kind} data-count={records.length}>
-      {kind === "handler" ? "Invocations panel" : "Executions panel"}
-    </div>
-  ),
-}));
-
-vi.mock("./execution-detail", () => ({
-  ExecutionDetailFetcher: (props: { executionId: string }) => (
-    <div data-testid="execution-detail-fetcher">{props.executionId}</div>
-  ),
-}));
-
-const mockNavigate = vi.fn();
-const mockCorrectUrl = vi.fn();
-
-vi.mock("wouter", () => createWouterMock({ useLocation: () => ["/apps/test_app/handlers", mockNavigate] }));
-
-vi.mock("../../hooks/use-correct-url", () => ({
-  useCorrectUrl: () => mockCorrectUrl,
-}));
-
-function renderHandlersTab(
-  listeners = [createListener({ listener_id: 1 })],
-  jobs = [createJob({ job_id: 10 })],
-  selectedHandler: string | null = null,
-) {
-  return renderWithAppState(
-    <HandlersTab
-      listeners={listeners}
-      jobs={jobs}
-      selectedHandler={selectedHandler}
-      selectedExecId={null}
-      appKey="test_app"
-    />,
-    { stateOverrides: { uptimeSeconds: signal<number | null>(120) } },
-  );
-}
-
-describe("HandlersTab", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it("renders the master list container", () => {
-    const { getByTestId } = renderHandlersTab();
-    expect(getByTestId("handler-list")).toBeDefined();
-  });
-
-  it("renders empty state when no listeners or jobs", () => {
-    const { getByTestId } = renderWithAppState(
-      <HandlersTab listeners={[]} jobs={[]} selectedHandler={null} selectedExecId={null} appKey="test_app" />,
-      { stateOverrides: { uptimeSeconds: signal<number | null>(120) } },
-    );
-    expect(getByTestId("handlers-empty")).toBeDefined();
-  });
-
-  it("shows detail pane placeholder when no item is selected", () => {
-    const { getByTestId } = renderHandlersTab();
-    expect(getByTestId("detail-placeholder")).toBeDefined();
-  });
-
-  it("shows listener detail pane when selectedHandler='listener/5'", () => {
-    const { getByTestId } = renderHandlersTab([createListener({ listener_id: 5 })], [], "listener/5");
-    expect(getByTestId("listener-detail-5")).toBeDefined();
-  });
-
-  it("shows job detail pane when selectedHandler='job/20'", () => {
-    const { getByTestId } = renderHandlersTab([], [createJob({ job_id: 20 })], "job/20");
-    expect(getByTestId("job-detail-20")).toBeDefined();
-  });
-
-  it("uses a semantic heading for the selected handler name", () => {
-    const listener = createListener({ listener_id: 6, handler_method: "on_garage_opened" });
-    const { getByRole } = renderHandlersTab([listener], [], "listener/6");
-    expect(getByRole("heading", { level: 2, name: "on_garage_opened" })).toBeDefined();
-  });
-
-  it("renders modifier chips for listener in detail pane", async () => {
-    const listener = createListener({
-      listener_id: 3,
-      debounce: 0.5,
-      throttle: null,
-      once: 0,
-      immediate: 0,
-    });
-    const { getByTestId, getByText } = renderHandlersTab([listener], [], "listener/3");
-    await waitFor(() => {
-      expect(getByTestId("listener-detail-3")).toBeDefined();
-    });
-    expect(getByText(/debounce/i)).toBeDefined();
-  });
-
-  it("renders schedule chips for job in detail pane", async () => {
-    const job = createJob({
-      job_id: 8,
-      trigger_label: "every 30s",
-      trigger_type: "Every",
-    });
-    const { getByTestId, getAllByText } = renderHandlersTab([], [job], "job/8");
-    await waitFor(() => {
-      expect(getByTestId("job-detail-8")).toBeDefined();
-    });
-    // "every 30s" appears in both the row description and the schedule chip
-    expect(getAllByText("every 30s").length).toBeGreaterThanOrEqual(1);
-  });
-
-  it("job detail: shows combined trigger label and detail in subtitle", async () => {
-    const job = createJob({
-      job_id: 9,
-      trigger_label: "every",
-      trigger_detail: "300s",
-      trigger_type: "interval",
-    });
-    const { getByTestId, getAllByText } = renderHandlersTab([], [job], "job/9");
-    await waitFor(() => {
-      expect(getByTestId("job-detail-9")).toBeDefined();
-    });
-    // Master row no longer renders a description line (see design/context.md two-line row layout);
-    // "every 5m" now only appears in the detail pane's schedule chip.
-    expect(getAllByText("every 5m").length).toBeGreaterThanOrEqual(1);
-  });
-
-  it("job detail: shows only detail when trigger_label is empty", async () => {
-    const job = createJob({
-      job_id: 10,
-      trigger_label: "",
-      trigger_detail: "300s",
-      trigger_type: "interval",
-    });
-    const { getByTestId, getAllByText } = renderHandlersTab([], [job], "job/10");
-    await waitFor(() => {
-      expect(getByTestId("job-detail-10")).toBeDefined();
-    });
-    expect(getAllByText("5m").length).toBeGreaterThanOrEqual(1);
-  });
-
-  it("does not show back button on desktop layout", () => {
-    const { queryByTestId } = renderHandlersTab();
-    expect(queryByTestId("back-to-list")).toBeNull();
-  });
-
-  it("handler detail: shows source location when available", async () => {
-    const listener = createListener({
-      listener_id: 8,
-      source_location: "garage_alerts.py:42",
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/8");
-    await waitFor(() => {
-      expect(getByTestId("handler-source-location")).toBeDefined();
-    });
-    expect(getByTestId("handler-source-location").textContent).toContain("garage_alerts.py");
-    expect(getByTestId("handler-source-location").textContent).toContain("42");
-  });
-
-  it("handler detail: stats row renders with counts", async () => {
-    const listener = createListener({
-      listener_id: 9,
-      total_invocations: 15,
-      failed: 2,
-      timed_out: 1,
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/9");
-    await waitFor(() => {
-      expect(getByTestId("handler-stats-row")).toBeDefined();
-    });
-    const statsRow = getByTestId("handler-stats-row");
-    expect(statsRow.textContent).toContain("15");
-    expect(statsRow.textContent).toContain("2");
-    expect(statsRow.textContent).toContain("1");
-  });
-
-  it("handler detail: shows error banner when listener has errors", async () => {
-    const listener = createListener({
-      listener_id: 11,
-      failed: 1,
-      last_error_type: "KeyError",
-      last_error_message: "missing key 'state'",
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/11");
-    await waitFor(() => {
-      expect(getByTestId("handler-error-banner")).toBeDefined();
-    });
-    expect(getByTestId("handler-error-banner").textContent).toContain("KeyError");
-  });
-
-  it("handler detail: shows registration source when available", async () => {
-    const listener = createListener({
-      listener_id: 12,
-      registration_source: "self.bus.on_state_change('light.kitchen', handler=self.on_light)",
-    });
-    const { getByRole, getByTestId, queryByTestId } = renderHandlersTab([listener], [], "listener/12");
-    await waitFor(() => {
-      expect(getByTestId("handler-registration-toggle")).toBeDefined();
-    });
-    expect(getByRole("region", { name: "Registration" })).toBeDefined();
-    const toggle = getByTestId("handler-registration-toggle");
-    expect(toggle.getAttribute("aria-controls")).toBe("listener-detail-12-registration-source-panel");
-    expect(toggle.textContent).toContain("show call");
-    // Registration is collapsed by default
-    expect(queryByTestId("handler-registration-source")).toBeNull();
-    fireEvent.click(toggle);
-    const registrationSource = getByTestId("handler-registration-source");
-    expect(registrationSource.id).toBe(toggle.getAttribute("aria-controls"));
-    expect(registrationSource.textContent).toContain("on_state_change");
-  });
-
-  it("handler detail: omits registration source when null", async () => {
-    const listener = createListener({ listener_id: 13, registration_source: null });
-    const { getByTestId, queryByTestId } = renderHandlersTab([listener], [], "listener/13");
-    await waitFor(() => {
-      expect(getByTestId("listener-detail-13")).toBeDefined();
-    });
-    expect(queryByTestId("handler-registration-source")).toBeNull();
-  });
-
-  it("handler stats row: renders err rate cell", async () => {
-    const listener = createListener({
-      listener_id: 20,
-      total_invocations: 10,
-      successful: 8,
-      failed: 2,
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/20");
-    await waitFor(() => {
-      expect(getByTestId("handler-stats-row")).toBeDefined();
-    });
-    const statsRow = getByTestId("handler-stats-row");
-    expect(statsRow.textContent).toContain("Err %");
-    expect(statsRow.textContent).toContain("20%");
-  });
-
-  it("handler stats row: does not show cancelled when zero", async () => {
-    const listener = createListener({ listener_id: 21, cancelled: 0 });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/21");
-    await waitFor(() => getByTestId("handler-stats-row"));
-    const statsRow = getByTestId("handler-stats-row");
-    expect(statsRow.textContent).not.toContain("Cancelled");
-  });
-
-  it("handler stats row: shows cancelled count when > 0", async () => {
-    const listener = createListener({ listener_id: 22, cancelled: 3 });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/22");
-    await waitFor(() => getByTestId("handler-stats-row"));
-    const statsRow = getByTestId("handler-stats-row");
-    expect(statsRow.textContent).toContain("Cancelled");
-    expect(statsRow.textContent).toContain("3");
-  });
-
-  it("handler stats row: shows Backpressure Dropped with warn tone when count > 0", async () => {
-    const listener = createListener({
-      listener_id: 24,
-      backpressure: "drop_newest",
-      backpressure_dropped_count: 5,
-      total_invocations: 5,
-      failed: 0,
-      timed_out: 0,
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/24");
-    await waitFor(() => getByTestId("handler-stats-row"));
-    const statsRow = getByTestId("handler-stats-row");
-    expect(statsRow.textContent).toContain("Backpressure Dropped");
-    // Dropped events are data loss: warns, with drop-rate percentage
-    const warnValue = statsRow.querySelector("[data-tone='warn']");
-    expect(warnValue?.textContent).toBe("5 (50%)");
-  });
-
-  it("handler stats row: shows 100% rate when every event was dropped", async () => {
-    const listener = createListener({
-      listener_id: 25,
-      backpressure: "drop_newest",
-      backpressure_dropped_count: 5,
-      total_invocations: 0,
-      failed: 0,
-      timed_out: 0,
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/25");
-    await waitFor(() => getByTestId("handler-stats-row"));
-    const statsRow = getByTestId("handler-stats-row");
-    const warnValue = statsRow.querySelector("[data-tone='warn']");
-    expect(warnValue?.textContent).toBe("5 (100%)");
-  });
-
-  it("handler stats row: shows — for avg when there are no executions", async () => {
-    const listener = createListener({
-      listener_id: 23,
-      min_duration_ms: null,
-      max_duration_ms: null,
-      avg_duration_ms: 0,
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/23");
-    await waitFor(() => getByTestId("handler-stats-row"));
-    const statsRow = getByTestId("handler-stats-row");
-    const cells = statsRow.querySelectorAll("[data-testid='handler-stats-row-cell']");
-    const avgCell = Array.from(cells).find((c) => c.textContent?.includes("Avg"));
-    expect(avgCell?.textContent).toContain("—");
-  });
-
-  it("handler error banner: shows expandable traceback when available", async () => {
-    const listener = createListener({
-      listener_id: 24,
-      failed: 1,
-      last_error_type: "ValueError",
-      last_error_message: "bad value",
-      last_error_traceback: "Traceback (most recent call last):\n  File test.py line 1\nValueError: bad value",
-    });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/24");
-    await waitFor(() => getByTestId("handler-error-banner"));
-    const banner = getByTestId("handler-error-banner");
-    // Traceback toggle button should be present
-    const toggle = banner.querySelector("[data-testid='traceback-toggle']");
-    expect(toggle).not.toBeNull();
-    // Traceback is initially collapsed
-    const tracebackContent = banner.querySelector("[data-testid='traceback-content']");
-    expect(tracebackContent).not.toBeNull();
-    // Expand it
-    fireEvent.click(toggle!);
-    expect(banner.textContent).toContain("Traceback (most recent call last)");
-  });
-
-  it("job detail: shows error banner when job has errors", async () => {
-    const job = createJob({
-      job_id: 30,
-      failed: 1,
-      last_error_type: "RuntimeError",
-      last_error_message: "something failed",
-      last_error_traceback: "Traceback (most recent call last):\nRuntimeError: something failed",
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/30");
-    await waitFor(() => getByTestId("job-detail-30"));
-    const banner = getByTestId("job-error-banner");
-    expect(banner.textContent).toContain("RuntimeError");
-    expect(banner.textContent).toContain("something failed");
-    // Toggle and check traceback
-    const toggle = banner.querySelector("[data-testid='traceback-toggle']");
-    expect(toggle).not.toBeNull();
-    fireEvent.click(toggle!);
-    expect(banner.textContent).toContain("Traceback (most recent call last)");
-  });
-
-  it("job stats row: renders err rate cell", async () => {
-    const job = createJob({
-      job_id: 31,
-      total_executions: 10,
-      successful: 7,
-      failed: 2,
-      timed_out: 1,
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/31");
-    await waitFor(() => getByTestId("job-stats-row"));
-    const statsRow = getByTestId("job-stats-row");
-    expect(statsRow.textContent).toContain("Err %");
-    expect(statsRow.textContent).toContain("20%");
-  });
-
-  it("job stats row: visually separates failed and timed_out as distinct cells", async () => {
-    const job = createJob({
-      job_id: 32,
-      failed: 2,
-      timed_out: 1,
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/32");
-    await waitFor(() => getByTestId("job-stats-row"));
-    const statsRow = getByTestId("job-stats-row");
-    // Both labels must be present as distinct cells
-    expect(statsRow.textContent).toContain("Failed");
-    expect(statsRow.textContent).toContain("Timed Out");
-    // Failed uses err color class, Timed Out uses warn color class
-    const errValue = statsRow.querySelector("[data-tone='err']");
-    const warnValue = statsRow.querySelector("[data-tone='warn']");
-    expect(errValue).not.toBeNull();
-    expect(warnValue).not.toBeNull();
-    expect(errValue?.textContent).toBe("2");
-    expect(warnValue?.textContent).toBe("1");
-  });
-
-  // URL-driven selection tests (T03)
-  it("selects listener by selectedHandler='listener/1' prop", () => {
-    const listeners = [createListener({ listener_id: 1 })];
-    const { getByTestId } = renderHandlersTab(listeners, [], "listener/1");
-    expect(getByTestId("listener-detail-1")).toBeDefined();
-  });
-
-  it("selects job by selectedHandler='job/10' prop", () => {
-    const jobs = [createJob({ job_id: 10 })];
-    const { getByTestId } = renderHandlersTab([], jobs, "job/10");
-    expect(getByTestId("job-detail-10")).toBeDefined();
-  });
-
-  it("shows detail placeholder when selectedHandler is null", () => {
-    const { getByTestId } = renderHandlersTab([createListener({ listener_id: 1 })], [], null);
-    expect(getByTestId("detail-placeholder")).toBeDefined();
-  });
-
-  it("calls correctUrl when selectedHandler references a non-existent listener", () => {
-    const listeners = [createListener({ listener_id: 1 })];
-    renderHandlersTab(listeners, [], "listener/999");
-    expect(mockCorrectUrl).toHaveBeenCalledWith("/apps/test_app/handlers");
-  });
-
-  it("calls correctUrl when selectedHandler references a non-existent job", () => {
-    const jobs = [createJob({ job_id: 1 })];
-    renderHandlersTab([], jobs, "job/999");
-    expect(mockCorrectUrl).toHaveBeenCalledWith("/apps/test_app/handlers");
-  });
-
-  it("does not call correctUrl when data is empty (loading guard)", () => {
-    // Empty arrays = loading state / no data — should not correct URL
-    renderHandlersTab([], [], "listener/999");
-    // The empty-state branch renders, no correctUrl call
-    expect(mockCorrectUrl).not.toHaveBeenCalled();
-  });
-
-  it("clicking a listener row navigates to handler deep-link URL", () => {
-    const listeners = [createListener({ listener_id: 5 })];
-    const { getByTestId } = renderHandlersTab(listeners, [], null);
-    fireEvent.click(getByTestId("unified-row-listener-5"));
-    expect(mockNavigate).toHaveBeenCalledWith("/apps/test_app/handlers/listener/5");
-  });
-
-  it("clicking a job row navigates to job deep-link URL", () => {
-    const jobs = [createJob({ job_id: 20 })];
-    const { getByTestId } = renderHandlersTab([], jobs, null);
-    fireEvent.click(getByTestId("unified-row-job-20"));
-    expect(mockNavigate).toHaveBeenCalledWith("/apps/test_app/handlers/job/20");
-  });
-
-  it("clicking a listener row includes instanceQs in deep-link URL", () => {
-    const listeners = [createListener({ listener_id: 3 })];
-    const { getByTestId } = renderWithAppState(
-      <HandlersTab
-        listeners={listeners}
-        jobs={[]}
-        selectedHandler={null}
-        selectedExecId={null}
-        appKey="test_app"
-        instanceIndex={1}
-      />,
-      { stateOverrides: { uptimeSeconds: signal<number | null>(120) } },
-    );
-    fireEvent.click(getByTestId("unified-row-listener-3"));
-    expect(mockNavigate).toHaveBeenCalledWith("/apps/test_app/handlers/listener/3?instance=1");
-  });
-
-  it("job detail: renders schedule chips when jitter and group are set", async () => {
-    const job = createJob({
-      job_id: 40,
-      jitter: 5,
-      group: "my-group",
-      trigger_type: "Every",
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/40");
-    await waitFor(() => getByTestId("job-detail-40"));
-    const chips = getByTestId("schedule-chips");
-    expect(chips.textContent).toContain("±5s jitter");
-    expect(chips.textContent).toContain("group: my-group");
-  });
-
-  it("job detail: shows next-run text in stats when next_run is set", async () => {
-    const job = createJob({
-      job_id: 42,
-      next_run: Date.now() / 1000 + 300,
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/42");
-    await waitFor(() => getByTestId("job-detail-42"));
-    const statsRow = getByTestId("job-stats-row");
-    expect(statsRow.textContent).toContain("Next");
-    expect(statsRow.textContent).toContain("next");
-  });
-
-  it("job detail: shows fire-at text in stats when fire_at is set but next_run is null", async () => {
-    const job = createJob({
-      job_id: 43,
-      next_run: null,
-      fire_at: Date.now() / 1000 + 60,
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/43");
-    await waitFor(() => getByTestId("job-detail-43"));
-    const statsRow = getByTestId("job-stats-row");
-    expect(statsRow.textContent).toContain("Next");
-    expect(statsRow.textContent).toContain("fire at");
-  });
-
-  it("job detail: shows failing badge when job has errors", async () => {
-    const job = createJob({
-      job_id: 44,
-      failed: 1,
-      last_error_type: "RuntimeError",
-      last_error_message: "boom",
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/44");
-    await waitFor(() => getByTestId("job-detail-44"));
-    expect(getByTestId("handler-status-pill").textContent).toBe("failing");
-  });
-
-  it("job detail: shows mode chip for every job", async () => {
-    const job = createJob({ job_id: 50, mode: "queued" });
-    const { getByTestId } = renderHandlersTab([], [job], "job/50");
-    await waitFor(() => getByTestId("job-detail-50"));
-    const modeChip = getByTestId("handler-mode-chip");
-    expect(modeChip.textContent).toBe("mode: queued");
-    expect(modeChip.getAttribute("data-variant")).toBe("muted");
-  });
-
-  it("job stats row: does not show Suppressed or Dropped when counts are zero", async () => {
-    const job = createJob({ job_id: 51, suppressed_count: 0, dropped_count: 0 });
-    const { getByTestId, queryByText } = renderHandlersTab([], [job], "job/51");
-    await waitFor(() => getByTestId("job-stats-row"));
-    expect(queryByText("Suppressed")).toBeNull();
-    expect(queryByText("Dropped")).toBeNull();
-  });
-
-  it("job stats row: shows Suppressed when suppressed_count > 0", async () => {
-    const job = createJob({ job_id: 52, mode: "single", suppressed_count: 3, dropped_count: 0 });
-    const { getByTestId, queryByText } = renderHandlersTab([], [job], "job/52");
-    await waitFor(() => getByTestId("job-stats-row"));
-    const statsRow = getByTestId("job-stats-row");
-    expect(statsRow.textContent).toContain("Suppressed");
-    expect(statsRow.textContent).toContain("3");
-    expect(queryByText("Dropped")).toBeNull();
-    // Suppressed is expected single-mode behavior: muted, not a warning
-    const muteValue = statsRow.querySelector("[data-tone='mute']");
-    expect(muteValue?.textContent).toBe("3");
-  });
-
-  it("job stats row: shows Dropped when dropped_count > 0", async () => {
-    const job = createJob({ job_id: 53, mode: "queued", suppressed_count: 0, dropped_count: 2 });
-    const { getByTestId, queryByText } = renderHandlersTab([], [job], "job/53");
-    await waitFor(() => getByTestId("job-stats-row"));
-    const statsRow = getByTestId("job-stats-row");
-    expect(statsRow.textContent).toContain("Dropped");
-    expect(statsRow.textContent).toContain("2");
-    expect(queryByText("Suppressed")).toBeNull();
-    // Dropped is data loss: warns
-    const warnValue = statsRow.querySelector("[data-tone='warn']");
-    expect(warnValue?.textContent).toBe("2");
-  });
-
-  it("job stats row: shows Skipped when skipped > 0", async () => {
-    const job = createJob({ job_id: 54, skipped: 4 });
-    const { getByTestId } = renderHandlersTab([], [job], "job/54");
-    await waitFor(() => getByTestId("job-stats-row"));
-    const statsRow = getByTestId("job-stats-row");
-    expect(statsRow.textContent).toContain("Skipped");
-    expect(statsRow.textContent).toContain("4");
-    const muteValue = statsRow.querySelector("[data-tone='mute']");
-    expect(muteValue?.textContent).toBe("4");
-  });
-
-  it("job stats row: does not show Skipped when count is zero", async () => {
-    const job = createJob({ job_id: 55, skipped: 0 });
-    const { getByTestId, queryByText } = renderHandlersTab([], [job], "job/55");
-    await waitFor(() => getByTestId("job-stats-row"));
-    expect(queryByText("Skipped")).toBeNull();
-  });
-
-  it("job detail: shows human_description as predicate description when present", async () => {
-    const job = createJob({
-      job_id: 56,
-      predicate_description: "<lambda>",
-      human_description: "binary_sensor.home_occupied is on",
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/56");
-    await waitFor(() => getByTestId("job-detail-56"));
-    expect(getByTestId("job-predicate-description").textContent).toBe("binary_sensor.home_occupied is on");
-  });
-
-  it("job detail: falls back to predicate_description when human_description is null", async () => {
-    const job = createJob({
-      job_id: 57,
-      predicate_description: "<lambda>",
-      human_description: null,
-    });
-    const { getByTestId } = renderHandlersTab([], [job], "job/57");
-    await waitFor(() => getByTestId("job-detail-57"));
-    expect(getByTestId("job-predicate-description").textContent).toBe("<lambda>");
-  });
-
-  it("job detail: does not show predicate description when both fields are null", async () => {
-    const job = createJob({ job_id: 58, predicate_description: null, human_description: null });
-    const { getByTestId, queryByTestId } = renderHandlersTab([], [job], "job/58");
-    await waitFor(() => getByTestId("job-detail-58"));
-    expect(queryByTestId("job-predicate-description")).toBeNull();
-  });
-
-  it("shows placeholder when selectedHandler has invalid format", () => {
-    const { queryByTestId, getByTestId } = renderHandlersTab(
-      [createListener({ listener_id: 1 })],
-      [],
-      "invalid-format",
-    );
-    expect(queryByTestId("listener-detail-1")).toBeNull();
-    expect(getByTestId("detail-placeholder")).toBeDefined();
-  });
-
-  it("handler detail: calls onSwitchToCode with line number when view-in-code clicked", async () => {
-    const onSwitch = vi.fn();
-    const listener = createListener({
-      listener_id: 45,
-      source_location: "my_app.py:99",
-    });
-    const { getByTestId } = renderWithAppState(
-      <HandlersTab
-        listeners={[listener]}
-        jobs={[]}
-        selectedHandler="listener/45"
-        selectedExecId={null}
-        appKey="test_app"
-        onSwitchToCode={onSwitch}
-      />,
-      { stateOverrides: { uptimeSeconds: signal<number | null>(120) } },
-    );
-    await waitFor(() => getByTestId("listener-detail-45"));
-    fireEvent.click(getByTestId("view-in-code-btn"));
-    expect(onSwitch).toHaveBeenCalledWith(99);
-  });
-
-  describe("job detail: Run Now button", () => {
-    it("renders in the job detail panel", async () => {
-      const job = createJob({ job_id: 60 });
-      const { getByTestId } = renderHandlersTab([], [job], "job/60");
-      await waitFor(() => getByTestId("job-detail-60"));
-      expect(getByTestId("run-now-btn")).toBeDefined();
-      expect(getByTestId("run-now-btn").textContent).toContain("Run Now");
-    });
-
-    it("enters loading state and disables on click", async () => {
-      const job = createJob({ job_id: 61 });
-      server.use(
-        http.post(
-          "/api/scheduler/jobs/:id/trigger",
-          () => new Promise(() => {}), // never resolves — keeps the button in loading state
-        ),
-      );
-      const { getByTestId } = renderHandlersTab([], [job], "job/61");
-      await waitFor(() => getByTestId("job-detail-61"));
-      const button = getByTestId("run-now-btn") as HTMLButtonElement;
-      fireEvent.click(button);
-      await waitFor(() => {
-        expect(button.disabled).toBe(true);
-      });
-    });
-
-    it("shows inline error on 409 response", async () => {
-      const job = createJob({ job_id: 62 });
-      server.use(
-        http.post("/api/scheduler/jobs/:id/trigger", () => {
-          return HttpResponse.json({ detail: "job is currently executing" }, { status: 409 });
-        }),
-      );
-      const { getByTestId } = renderHandlersTab([], [job], "job/62");
-      await waitFor(() => getByTestId("job-detail-62"));
-      fireEvent.click(getByTestId("run-now-btn"));
-      await waitFor(() => {
-        expect(getByTestId("run-now-error").textContent).toContain("job is currently executing");
-      });
-    });
-
-    it("re-enables after the request completes", async () => {
-      const job = createJob({ job_id: 63 });
-      const { getByTestId } = renderHandlersTab([], [job], "job/63");
-      await waitFor(() => getByTestId("job-detail-63"));
-      const button = getByTestId("run-now-btn") as HTMLButtonElement;
-      fireEvent.click(button);
-      await waitFor(() => {
-        expect(button.disabled).toBe(false);
-      });
-    });
-  });
-
-  it("listener detail: shows mode chip", async () => {
-    const listener = createListener({ listener_id: 70, mode: "queued" });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/70");
-    await waitFor(() => getByTestId("listener-detail-70"));
-    const modeChip = getByTestId("handler-mode-chip");
-    expect(modeChip.textContent).toBe("mode: queued");
-    expect(modeChip.getAttribute("data-variant")).toBe("muted");
-  });
-
-  it("handler detail: registration source stays hidden until the toggle is clicked", async () => {
-    const listener = createListener({
-      listener_id: 71,
-      registration_source: "self.bus.on_state_change('light.kitchen', handler=self.on_light)",
-    });
-    const { getByTestId, queryByTestId } = renderHandlersTab([listener], [], "listener/71");
-    await waitFor(() => getByTestId("listener-detail-71"));
-    expect(queryByTestId("handler-registration-source")).toBeNull();
-    const toggle = getByTestId("handler-registration-toggle");
-    expect(toggle.textContent).toContain("show call");
-    fireEvent.click(toggle);
-    expect(getByTestId("handler-registration-source")).toBeDefined();
-    expect(toggle.textContent).toContain("hide call");
-  });
-
-  it("handler detail: renders the executions table inside the detail card", async () => {
-    const listener = createListener({ listener_id: 72 });
-    const { getByTestId } = renderHandlersTab([listener], [], "listener/72");
-    await waitFor(() => getByTestId("listener-detail-72"));
-    const table = await waitFor(() => getByTestId("invocation-table-72"));
-    const detailCard = getByTestId("listener-detail-72");
-    expect(detailCard.contains(table)).toBe(true);
-  });
-
-  it("renders execution detail even when no handlers are registered", () => {
-    const { getByTestId, queryByTestId } = renderWithAppState(
-      <HandlersTab listeners={[]} jobs={[]} selectedHandler="listener/5" selectedExecId="abc-123" appKey="test_app" />,
-      { stateOverrides: { uptimeSeconds: signal<number | null>(120) } },
-    );
-    expect(getByTestId("execution-detail-fetcher")).toBeTruthy();
-    expect(queryByTestId("handlers-empty")).toBeNull();
-  });
-});
